Remove usage of word 'monotonic' from counter tests because all counters are now monotonic
